+++ src/config.php
@@ -248,5 +248,19 @@ return [
 
         // The name of the JavaScript event that will trigger a script inject.
         //'injectScriptEvent' => 'DOMContentLoaded',
+
+        // The maximum length of URIs that may be cached. Increasing this value requires manually updating the limit in the `uri` column of the `blitz_caches` database table. Note that the prefix length limit is 3072 bytes for InnoDB tables that use the DYNAMIC or COMPRESSED row format. Assuming a `utf8mb4` character set and a maximum of 4 bytes for each character, this is 768 characters.
+        // https://dev.mysql.com/doc/refman/8.0/en/column-indexes.html#column-indexes-prefix
+        // Warning: if using the File Storage driver, this value should not exceed 255 unless using a file system that supports longer file names.
+        // https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits
+        //'maxUriLength' => 255,
+
+        // The maximum length of SSI values that may be cached.
+        // https://nginx.org/en/docs/http/ngx_http_ssi_module.html#ssi_value_length
+        //maxSsiValueLength => 256,
+
+        // The maximum length of ESI values that may be cached.
+        // https://nginx.org/en/docs/http/ngx_http_ssi_module.html#ssi_value_length
+        //maxEsiValueLength => 256,
     ],
 ];