+++ src/services/FlushCacheService.php
@@ -156,7 +156,7 @@ class FlushCacheService extends Component
         $sql = '';
 
         if ($dbDriver == 'mysql') {
-            $sql = 'ALTER TABLE ' . $table . ' AUTO_INCREMENT = 1';
+            $sql = 'ALTER TABLE ' . $table . ' AUTO_INCREMENT=1';
         } elseif ($dbDriver == 'postgres') {
             $sql = 'ALTER SEQUENCE ' . $table . '_id_seq RESTART WITH 1';
         }