+++ tests/unit/drivers/RedisStorageTest.php
@@ -7,6 +7,7 @@ namespace putyourlightson\blitztests\unit\drivers;
 
 use Codeception\Test\Unit;
 use Craft;
+use craft\helpers\StringHelper;
 use putyourlightson\blitz\Blitz;
 use putyourlightson\blitz\drivers\storage\CacheStorageInterface;
 use putyourlightson\blitz\drivers\storage\YiiCacheStorage;
@@ -41,7 +42,7 @@ class RedisStorageTest extends Unit
      */
     private string $output = 'xyz';
 
-    protected function _before()
+    protected function _before(): void
     {
         parent::_before();
 
@@ -74,8 +75,10 @@ class RedisStorageTest extends Unit
         Blitz::$plugin->cacheStorage->save($this->output, $this->siteUri);
     }
 
-    public function testSave()
+    public function testSaveLongUri()
     {
+        $this->siteUri->uri = StringHelper::randomString(1000);
+        Blitz::$plugin->cacheStorage->save($this->output, $this->siteUri);
         $value = $this->cacheStorage->get($this->siteUri);
         $this->assertStringContainsString($this->output, $value);
     }