+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added the ability to send compressed responses to browsers that accept supported encodings.
 - Added the ability to save compressed cached values in the Yii Cache Storage to help reduce the memory required.
 - Added tips that display whether `gzip` is enabled on the web server in the Cache Storage settings.
+- Added the `maxUriLength`, `maxSsiValueLength` and `maxEsiValueLength` config settings.
 
 ### Changed
 - Cached includes and pages that contain SSI or ESI includes are now never compressed.